Update point type to use i32

+++ 2017/20/solver.rs
@@ -37,7 +37,7 @@ impl Particle {
         }
     }
 
-    fn size(&self) -> i64 {
+    fn size(&self) -> i32 {
         self.position.length()
     }
 }
